Use Cranelift instead of LLVM
This was a lot of work for seemingly little benefit, so here's the
rationale (in vague order of importance):

- Cranelift is written in Rust, not C++, so there's no FFI
- Cranelift does not have segmentation faults or undefined behavior if
you use the crate improperly
- Cranelift has releases instead of running from a Git branch, so things
don't break unexpectedly (I had to redo a lot of code when Inkwell
removed `Void::ptr_type`).
- Cranelift can compile directly to an object file instead of IR (using
cranelift_faerie)
- Cranelift runs significantly faster, although it performs fewer
optimizations.

+++ src/data.rs
@@ -2,6 +2,8 @@ use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::fmt::{self, Display, Formatter};
 
+use crate::backend::SIZE_T;
+
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 pub enum Keyword {
     // keywords
@@ -398,7 +400,7 @@ pub enum LengthError {
 }
 
 impl ArrayType {
-    pub fn length(&self) -> Result<u32, LengthError> {
+    pub fn length(&self) -> Result<SIZE_T, LengthError> {
         match self {
             ArrayType::Unbounded => Err(LengthError::Unbounded),
             ArrayType::Fixed(expr) => {
